Mark some use-cases doc in need of refresh
If anyone thinks one of these does not need updates, then we can remove
it from this list. I added a notice to each doc that I haven't recently
reviewed, (regardless of whether I am the original author.)

The use-cases docs that are not covered here are all ones intended to be
migrated into the Installation guides, per the GSoD plan that is being
advanced here.

So the use-cases section of the docs may go away entirely, (unless some
of these docs turn out to survive all this without getting relocated):

https://github.com/fluxcd/website/issues/1560

Signed-off-by: Kingdon Barrett <kingdon@weave.works>